Version 2.0 for parser
Second iteration of parser. Can be updated with new actions and prepositions as we work more on project. For now only splits the input into commands, possible objects and relevant prepositions. It will be up to a separate method to verify if input makes sense that I didn't want to work on till I look at how the other parts are implemented. The temp_parser is hence a placeholder for a sample of how that logic would look like.