Adding talk handler
Added basic functionality for talk handler. So far, it allows you to talk to the fairy in the kitchen, who wants blackberries. The blackberries need to be used on the fairy in order to change the fairy item to the hidden object version of the fairy with the second description.