Update tba_parser.py
Added ghost and giant mushroom dialogue into the talk command. Also I wrote code for printing out ascii art when certain objects are examined into the examine_handler. I updated the take and eat commands to allow for blackberries to be taken or eaten without removing the blackberries from the room. This is because they are edible, so if the players ends up eating the blackberries, they should still be able to get blackberries again to give to the kitchen fairy.